Moves Paginator from `@/common` to `lib/components/common`
